Restore somewhat-correct bounding sphere calculation for three >=151 (#222)
### Public-Facing Changes

Restored the bounding sphere behavior for three.js versions >=r151.

### Description

InstancedMesh now supports bounding sphere calculation, but it uses `instanceMatrix` which we do not set in this code. Instead, fall back to the geometry's boundingSphere, which is what it used to do before r151. This only uses the non-instanced vertices, but the Label's `position` still gets correctly applied which makes depth sorting mostly work (the same way it used to).